file rename bugfix
